- removed unused ASCIICast logger - merged logger.go into config.go
