+++ ansible/roles/ossh/templates/config.yaml.j2
@@ -6,7 +6,6 @@
 # For devs, enable to print debug messages 
 # for the respective subsystem.
 debug:
-  asciicast_v2: false
   fake_shell: false
   sync_commands: false
   sync_server: false
