+++ config.example.yaml
@@ -1,5 +1,4 @@
 debug:
-  asciicast_v2: false
   fake_shell: false
   sync_commands: false
   sync_server: false
